feat(sync): runSync только вручную; ночная сверка в кроне 04:35; зависшие running → error
- крон 0 */3: runSync убран, syncGoodsCategories на 10 */3 (бёрст на :00 ронял
  «Заботу» лимитом в 12:00); 35 4 мск — reconcileDaily по салонам последовательно
- при старте closeStaleSyncRuns (на деве закрыто 37 строк)
- POST /api/sync → 409, пока идёт прогон
- ЖИВОЙ ПРОГОН scripts/yclients-reconcile-e2e.js нашёл два дефекта типов:
  yclients_card_balance numeric + bonus_balance integer одним параметром
  («inconsistent types») и дробный баланс карты 784.89 в linkClientCard без
  ::numeric («invalid input syntax for type integer») — второе живёт и на проде
  в пути начисления: клиент с дробным балансом карты не привязывался никогда.
  Итог прогона: records=79 clients=60 errors=0 за 64 с.

+++ backend/services/loyalty.js
@@ -417,7 +417,7 @@ async function runSync(salon, syncType, userId) {
               await db.query(
                 `UPDATE clients SET
                    yclients_card_id=$1, yclients_card_number=$2, yclients_card_balance=$3,
-                   bonus_balance=$4, updated_at=NOW()
+                   bonus_balance=$4::numeric, updated_at=NOW()
                  WHERE id=$5`,
                 [card.id, card.number || card.loyalty_card_number || null,
                  cardBalance, cardBalance, dbClientId]
@@ -587,9 +587,14 @@ async function linkClientCard(salon, client, settings) {
   const cardNumber  = card.number || card.loyalty_card_number || null;
   const level = settings?.levels ? getLevel(paidAmount, settings.levels) : null;
 
+  // bonus_balance — INTEGER, баланс карты бывает дробным (784.89): без явного
+  // ::numeric pg выводит тип параметра как integer и падает «invalid input syntax
+  // for type integer» — карта не привязывается НИКОГДА, а в начислении это
+  // «сбой YClients → повтор на следующем событии» по кругу (поймано живым
+  // прогоном ночной сверки 25.09; numeric→integer при присваивании округляется).
   return db.one(
     `UPDATE clients SET yclients_card_id=$1,yclients_card_number=$2,yclients_card_balance=$3,
-     bonus_balance=$4,total_spent=$5,visits_count=$6,loyalty_level=$7,updated_at=NOW()
+     bonus_balance=$4::numeric,total_spent=$5,visits_count=$6,loyalty_level=$7,updated_at=NOW()
      WHERE id=$8 RETURNING *`,
     [card.id, cardNumber, cardBalance, cardBalance, paidAmount, visitsCount,
      level?.key || client.loyalty_level, client.id]
